passing calcConfidence and cslParser tests
